+++ packages/react/src/components/UIShell/__stories__/UIShell.stories.js
@@ -9,11 +9,12 @@
 
 import React, { useState, useRef } from 'react';
 import mdx from './UIShell.mdx';
-import { SIDE_NAV_TYPE, SideNav } from '../components/SideNav';
+import { SideNav } from '../components/SideNav';
 import { SideNavItems } from '../components/SideNavItems';
 import { SideNavMenu } from '../components/SideNavMenu';
 import { SideNavMenuItem } from '../components/SideNavMenuItem';
 import { SideNavLink } from '../components/SideNavLink';
+import { SideNavSlot } from '../components/SideNavSlot';
 import {
   HeaderPopover,
   HeaderPopoverActions,
@@ -39,6 +40,7 @@ import {
   ExpandableSearch,
   Link,
   Button,
+  Dropdown,
 } from '@carbon/react';
 import {
   Add,
@@ -64,6 +66,8 @@ import {
   ShoppingCart,
   Share,
   User,
+  VirtualColumnKey,
+  Menu,
 } from '@carbon/icons-react';
 
 import {
@@ -90,6 +94,7 @@ export default {
     SideNavLink,
     SideNavMenu,
     SideNavMenuItem,
+    SideNavSlot,
     TrialCountdown,
   },
   parameters: {
@@ -328,64 +333,127 @@ export const Default = () => {
             onOverlayClick={onClickSideNavExpand}
             className="nav--global">
             <SideNavItems>
-              <SideNavMenu renderIcon={CarbonDesignSystem} title="Product 1">
-                <SideNavMenu renderIcon={Home} title="Home">
+              <SideNavMenu
+                renderIcon={CarbonDesignSystem}
+                title="Product 1"
+                primary
+                defaultExpanded>
+                <SideNavSlot renderIcon={VirtualColumnKey}>
+                  <Dropdown
+                    aria-label="Choose an option"
+                    id="default"
+                    size="sm"
+                    itemToString={(item) => (item ? item.text : '')}
+                    items={[
+                      { text: 'Option 1' },
+                      { text: 'Option 2' },
+                      { text: 'Option 3' },
+                    ]}
+                    label="Choose an option"
+                  />
+                </SideNavSlot>
+                <SideNavDivider />
+                <SideNavMenu renderIcon={Home} title="Home" defaultExpanded>
                   <SideNavMenuItem href="http://www.carbondesignsystem.com">
                     Item level 3
                   </SideNavMenuItem>
-                  <SideNavMenuItem href="http://www.carbondesignsystem.com">
+                  <SideNavMenuItem
+                    isActive
+                    href="http://www.carbondesignsystem.com">
                     Item level 3
                   </SideNavMenuItem>
                   <SideNavMenuItem href="http://www.carbondesignsystem.com">
                     Item level 3
                   </SideNavMenuItem>
                 </SideNavMenu>
-                <SideNavMenuItem
+                <SideNavLink
                   href="http://www.carbondesignsystem.com"
                   renderIcon={BusinessProcesses}>
                   Business
-                </SideNavMenuItem>
-                <SideNavMenuItem
+                </SideNavLink>
+                <SideNavLink
                   href="http://www.carbondesignsystem.com"
                   renderIcon={Application}>
                   Applications
-                </SideNavMenuItem>
-                <SideNavMenuItem
+                </SideNavLink>
+                <SideNavLink
                   href="http://www.carbondesignsystem.com"
                   renderIcon={Platforms}>
                   Platforms
-                </SideNavMenuItem>
-                <SideNavMenu renderIcon={Layers} title="Infrastructure">
-                  <SideNavMenuItem href="http://www.carbondesignsystem.com">
-                    Item level 3
-                  </SideNavMenuItem>
-                  <SideNavMenuItem href="http://www.carbondesignsystem.com">
-                    Item level 3
-                  </SideNavMenuItem>
-                  <SideNavMenuItem href="http://www.carbondesignsystem.com">
-                    Item level 3
-                  </SideNavMenuItem>
-                </SideNavMenu>
+                </SideNavLink>
+                <SideNavLink
+                  href="http://www.carbondesignsystem.com"
+                  renderIcon={Layers}>
+                  Infrastructure
+                </SideNavLink>
+                <SideNavDivider />
+                <SideNavLink
+                  href="http://www.carbondesignsystem.com"
+                  renderIcon={Dashboard}>
+                  Dashboard
+                </SideNavLink>
+                <SideNavLink
+                  href="http://www.carbondesignsystem.com"
+                  renderIcon={DataAnalytics}>
+                  Analytics
+                </SideNavLink>
+                <SideNavLink
+                  href="http://www.carbondesignsystem.com"
+                  renderIcon={EventIncident}>
+                  Incidents
+                </SideNavLink>
+                <SideNavLink
+                  href="http://www.carbondesignsystem.com"
+                  renderIcon={Security}>
+                  Security
+                </SideNavLink>
+                <SideNavLink
+                  href="http://www.carbondesignsystem.com"
+                  renderIcon={WorkflowAutomation}>
+                  Automations
+                </SideNavLink>
+                <SideNavDivider />
+                <SideNavLink
+                  href="http://www.carbondesignsystem.com"
+                  renderIcon={DocumentMultiple_01}>
+                  Docs
+                </SideNavLink>
+                <SideNavLink
+                  href="http://www.carbondesignsystem.com"
+                  renderIcon={Settings}>
+                  Settings
+                </SideNavLink>
+                <SideNavLink
+                  href="http://www.carbondesignsystem.com"
+                  renderIcon={OverflowMenuVertical}>
+                  More
+                </SideNavLink>
               </SideNavMenu>
-              <SideNavMenu renderIcon={CarbonIBMDotCom} title="Product 2">
+              <SideNavMenu
+                renderIcon={CarbonIBMDotCom}
+                title="Product 2"
+                primary>
                 <SideNavMenuItem
                   renderIcon={Home}
                   href="http://www.carbondesignsystem.com">
-                  Link
+                  Home product 2
                 </SideNavMenuItem>
               </SideNavMenu>
-              <SideNavMenu renderIcon={CarbonforIBMProducts} title="Product 3">
+              <SideNavMenu
+                renderIcon={CarbonforIBMProducts}
+                title="Product 3"
+                primary>
                 <SideNavMenuItem
                   renderIcon={Home}
                   href="http://www.carbondesignsystem.com">
-                  Link
+                  Home product 3
                 </SideNavMenuItem>
               </SideNavMenu>
-              <SideNavMenu renderIcon={IBMTelemetry} title="Product 4">
+              <SideNavMenu renderIcon={IBMTelemetry} title="Product 4" primary>
                 <SideNavMenuItem
                   renderIcon={Home}
                   href="http://www.carbondesignsystem.com">
-                  Link
+                  Home product 4
                 </SideNavMenuItem>
               </SideNavMenu>
               <SideNavDivider />
@@ -407,6 +475,23 @@ export const Default = () => {
             isRail
             aria-label="Product navigation">
             <SideNavItems>
+              <SideNavSlot renderIcon={Menu}>
+                <Menu />
+              </SideNavSlot>
+              <SideNavSlot renderIcon={VirtualColumnKey}>
+                <Dropdown
+                  id="default"
+                  size="sm"
+                  itemToString={(item) => (item ? item.text : '')}
+                  items={[
+                    { text: 'Option 1' },
+                    { text: 'Option 2' },
+                    { text: 'Option 3' },
+                  ]}
+                  label="Choose an option"
+                />
+              </SideNavSlot>
+              <SideNavDivider />
               <SideNavMenu renderIcon={Home} title="Home">
                 <SideNavMenuItem href="http://www.carbondesignsystem.com">
                   Item level 3
@@ -502,11 +587,26 @@ Default.parameters = {
  */
 export const SideNavStory = () => (
   <SideNav
+    isTreeview
     isFixedNav
     expanded
     isChildOfHeader={false}
     aria-label="Side navigation">
     <SideNavItems>
+      <SideNavSlot renderIcon={VirtualColumnKey}>
+        <Dropdown
+          id="default"
+          size="sm"
+          itemToString={(item) => (item ? item.text : '')}
+          items={[
+            { text: 'Option 1' },
+            { text: 'Option 2' },
+            { text: 'Option 3' },
+          ]}
+          label="Choose an option"
+        />
+      </SideNavSlot>
+      <SideNavDivider />
       <SideNavMenu renderIcon={Fade} title="Sub-menu level 1">
         <SideNavMenuItem href="http://www.carbondesignsystem.com">
           Link
@@ -560,6 +660,78 @@ export const SideNavStory = () => (
   </SideNav>
 );
 SideNavStory.storyName = 'SideNav';
+
+/**
+ * Story for SideNavDoublewide
+ * @returns {React.ReactElement} The JSX for the story
+ */
+export const SideNavDoubleWideStory = () => (
+  <SideNav
+    isTreeview
+    isFixedNav
+    expanded
+    isChildOfHeader={false}
+    aria-label="Side navigation">
+    <SideNavItems>
+      <SideNavMenu
+        renderIcon={Fade}
+        title="Sub-menu level 1"
+        primary
+        defaultExpanded>
+        <SideNavSlot>
+          <Dropdown
+            aria-label="Choose an option"
+            id="default"
+            size="sm"
+            itemToString={(item) => (item ? item.text : '')}
+            items={[
+              { text: 'Option 1' },
+              { text: 'Option 2' },
+              { text: 'Option 3' },
+            ]}
+            label="Choose an option"
+          />
+        </SideNavSlot>
+        <SideNavDivider />
+        <SideNavMenu renderIcon={Fade} title="Sub-menu level 2">
+          <SideNavMenuItem href="http://www.carbondesignsystem.com">
+            Item level 3
+          </SideNavMenuItem>
+        </SideNavMenu>
+        <SideNavMenuItem
+          renderIcon={Fade}
+          href="http://www.carbondesignsystem.com">
+          Item level 2
+        </SideNavMenuItem>
+        <SideNavMenuItem
+          renderIcon={Fade}
+          href="http://www.carbondesignsystem.com">
+          Item level 2
+        </SideNavMenuItem>
+        <SideNavMenuItem
+          renderIcon={Fade}
+          href="http://www.carbondesignsystem.com">
+          Item level 2
+        </SideNavMenuItem>
+      </SideNavMenu>
+      <SideNavMenu renderIcon={Fade} title="Sub-menu level 1" primary>
+        <SideNavMenuItem
+          renderIcon={Fade}
+          href="http://www.carbondesignsystem.com">
+          Item level 2
+        </SideNavMenuItem>
+      </SideNavMenu>
+      <SideNavDivider />
+      <SideNavLink renderIcon={Fade} href="http://www.carbondesignsystem.com">
+        Link level 1
+      </SideNavLink>
+      <SideNavLink renderIcon={Fade} href="http://www.carbondesignsystem.com">
+        Link level 1
+      </SideNavLink>
+    </SideNavItems>
+  </SideNav>
+);
+SideNavDoubleWideStory.storyName = 'SideNav with Double Wide';
 
 /**
  * Story for SideNav w/TreeView
@@ -655,6 +827,21 @@ export const SideNavRail = () => (
     isChildOfHeader={false}
     aria-label="Product navigation">
     <SideNavItems>
+      <SideNavSlot renderIcon={VirtualColumnKey}>
+        <Dropdown
+          aria-label="Choose an option"
+          id="default"
+          size="sm"
+          itemToString={(item) => (item ? item.text : '')}
+          items={[
+            { text: 'Option 1' },
+            { text: 'Option 2' },
+            { text: 'Option 3' },
+          ]}
+          label="Choose an option"
+        />
+      </SideNavSlot>
+      <SideNavDivider />
       <SideNavMenu renderIcon={Fade} title="Sub-menu level 1">
         <SideNavMenuItem href="http://www.carbondesignsystem.com">
           Item level 2
@@ -721,120 +908,6 @@ export const SideNavRail = () => (
       </SideNavLink>
     </SideNavItems>
   </SideNav>
-);
-
-/**
- * Story for SideNav panel
- * @returns {React.ReactElement} The JSX for the story
- */
-export const SideNavPanel = () => (
-  <>
-    <SideNav
-      navType={SIDE_NAV_TYPE.PANEL}
-      isChildOfHeader={false}
-      hideOverlay
-      aria-label="Product navigation">
-      <SideNavItems>
-        <SideNavMenu renderIcon={Fade} title="Sub-menu level 1">
-          <SideNavMenuItem href="http://www.carbondesignsystem.com">
-            Item level 2
-          </SideNavMenuItem>
-          <SideNavMenuItem href="http://www.carbondesignsystem.com">
-            Item level 2
-          </SideNavMenuItem>
-          <SideNavMenuItem href="http://www.carbondesignsystem.com">
-            Item level 2
-          </SideNavMenuItem>
-        </SideNavMenu>
-        <SideNavMenu renderIcon={Fade} title="Sub-menu level 1">
-          <SideNavMenuItem href="http://www.carbondesignsystem.com">
-            Item level 2
-          </SideNavMenuItem>
-          <SideNavMenuItem href="http://www.carbondesignsystem.com">
-            Item level 2
-          </SideNavMenuItem>
-          <SideNavMenuItem href="http://www.carbondesignsystem.com">
-            Item level 2
-          </SideNavMenuItem>
-        </SideNavMenu>
-        <SideNavMenu renderIcon={Fade} title="Sub-menu level 1">
-          <SideNavMenuItem href="http://www.carbondesignsystem.com">
-            Item level 2
-          </SideNavMenuItem>
-          <SideNavMenuItem href="http://www.carbondesignsystem.com">
-            Item level 2
-          </SideNavMenuItem>
-          <SideNavMenuItem href="http://www.carbondesignsystem.com">
-            Item level 2
-          </SideNavMenuItem>
-        </SideNavMenu>
-        <SideNavMenu renderIcon={Fade} title="Sub-menu level 1">
-          <SideNavMenuItem href="http://www.carbondesignsystem.com">
-            Item level 2
-          </SideNavMenuItem>
-          <SideNavMenuItem href="http://www.carbondesignsystem.com">
-            Item level 2
-          </SideNavMenuItem>
-          <SideNavMenuItem href="http://www.carbondesignsystem.com">
-            Item level 2
-          </SideNavMenuItem>
-        </SideNavMenu>
-        <SideNavDivider />
-        <SideNavLink renderIcon={Fade} href="http://www.carbondesignsystem.com">
-          Link
-        </SideNavLink>
-        <SideNavLink renderIcon={Fade} href="http://www.carbondesignsystem.com">
-          Link
-        </SideNavLink>
-        <SideNavLink renderIcon={Fade} href="http://www.carbondesignsystem.com">
-          Link
-        </SideNavLink>
-        <SideNavLink renderIcon={Fade} href="http://www.carbondesignsystem.com">
-          Link
-        </SideNavLink>
-        <SideNavDivider />
-        <SideNavLink renderIcon={Fade} href="http://www.carbondesignsystem.com">
-          Link
-        </SideNavLink>
-        <SideNavLink renderIcon={Fade} href="http://www.carbondesignsystem.com">
-          Link
-        </SideNavLink>
-      </SideNavItems>
-    </SideNav>
-    <Content>
-      <Grid align="start">
-        <Column sm={4} md={8} lg={12}>
-          <h2 style={{ margin: '0 0 30px 0' }}>Purpose and function</h2>
-          <p>
-            The shell is perhaps the most crucial piece of any UI built with{' '}
-            {''}
-            <a href="www.carbondesignsystem.com">Carbon</a>. It contains the
-            shared navigation framework for the entire design system and ties
-            the products in IBM’s portfolio together in a cohesive and elegant
-            way. The shell is the home of the topmost navigation, where users
-            can quickly and dependably gain their bearings and move between
-            pages.
-            <br />
-            <br />
-            The shell was designed with maximum flexibility built in, to serve
-            the needs of a broad range of products and users. Adopting the shell
-            ensures compliance with IBM design standards, simplifies development
-            efforts, and provides great user experiences. All IBM products built
-            with Carbon are required to use the shell’s header.
-            <br />
-            <br />
-            To better understand the purpose and function of the UI shell,
-            consider the “shell” of MacOS, which contains the Apple menu,
-            top-level navigation, and universal, OS-level controls at the top of
-            the screen, as well as a universal dock along the bottom or side of
-            the screen. The Carbon UI shell is roughly analogous in function to
-            these parts of the Mac UI. For example, the app switcher portion of
-            the shell can be compared to the dock in MacOS.
-          </p>
-        </Column>
-      </Grid>
-    </Content>
-  </>
 );
 
 /**